moves external URL status tests into playwright
