+++ frontend/package.json
@@ -55,7 +55,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "e2e": "npx playwright install && PWDEBUG=console npx playwright test",
+    "e2e": "npx playwright install && npx playwright test",
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
