+++ frontend/playwright.config.ts
@@ -30,6 +30,8 @@ const config: PlaywrightTestConfig = {
   },
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
+  /* run all tests, even those within a shared file, to run in parallel  */
+  fullyParallel: true,
   /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
   /* Opt out of parallel tests on CI. */
