+++ frontend/src/pages/AboutUs/OurTeamTab.test.ts
@@ -2,13 +2,8 @@
  * @jest-environment node
  */
 
-import { GOOGLE_FELLOWS, PARTNERS } from "./OurTeamTab";
-import {
-  WAIT_FOR_URL_STATUSES,
-  SUCCESS_CODE,
-  getTestableUrls,
-  getStatus,
-} from "../../utils/externalUrls.test";
+import { GOOGLE_FELLOWS, PARTNERS } from "./OurTeamData";
+import { getTestableUrls } from "../../utils/externalUrls.test";
 
 // Collect all OurTeam URLS
 const allUrls: string[] = [];
@@ -22,15 +17,7 @@ for (const partner of PARTNERS) {
 const testUrls = getTestableUrls(allUrls);
 
 describe.each(testUrls)("OurTeamTab URL %s", (testUrl) => {
-  test.concurrent(
-    `should return 200 (SUCCESS)`,
-    async () => {
-      const urlStatus = await getStatus(testUrl);
-      expect(urlStatus).toEqual(SUCCESS_CODE);
-    },
-    WAIT_FOR_URL_STATUSES
-  );
   test("should use HTTPS", () => {
-    expect(testUrl.slice(0, 8)).toEqual("https://");
+    expect(testUrl.startsWith("https://")).toBe(true);
   });
 });
